Revert some changes
